Ver 0.3.0 - Verilog Modules!
Summary :
Will now also spit out a verilog module corresponding to the logic
interpreted in cout!

interpreter:
Moved #include of the header file into ifndef, not sure why it was
outside to begin with
Added tracking of redstone power level to cut off recursive dust
searches once the distance is too far for any power source to power the
current component
Fixed error in rm_dup_relationship that will remove the wrong lines when
found a duplicate
Fixed error in rm_dup_relationship where after deleting a relationship,
the new relationship that then moves down into the spot of the recently
deleted will not be checked whether it is a duplicate

main :
Added lines to output to cout the Verilog module created

Makefile:
Fixed error with where LFLAGS were included in the commands
Added commands for compiling toVer

toVer:
These files are for taking a relationship table and creating a Verilog
module from the relationship table. The notable functions in these files
is create_module, which will output to cout a Verilog module of the
relationship table fed into it.